Change max precision
Magento stores float eav attributes with up to 4 decimals

+++ app/code/community/Aurmil/SetPricePrecision/Model/Directory/Currency.php
@@ -5,6 +5,6 @@ extends Mage_Directory_Model_Currency
 {
     public function format($price, $options=array(), $includeContainer = true, $addBrackets = false)
     {
-        return $this->formatPrecision($price, Mage::getStoreConfig('catalog/price/precision'), $options, $includeContainer, $addBrackets);
+        return $this->formatPrecision($price, (int)Mage::getStoreConfig('catalog/price/precision'), $options, $includeContainer, $addBrackets);
     }
 }
